added multiple args check
