+++ inc/command_parsing.h
@@ -6,7 +6,7 @@
 /*   By: mdomnik <mdomnik@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/27 02:18:21 by mdomnik           #+#    #+#             */
-/*   Updated: 2024/08/27 03:39:52 by mdomnik          ###   ########.fr       */
+/*   Updated: 2024/08/27 03:49:23 by mdomnik          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 # define COMMAND_PARSING_H
 
 //file_parser.c
-int	validate_arg_count(int argc);
+int	check_arg_count(char **argv);
 
 //find_scene_file.c
 int	validate_file_name(char **argv, t_options *options);
@@ -22,6 +22,5 @@ int	scene_file_check(char *filename, char **scene_file, bool *scene_count);
 int	check_scene_file(char *args);
 
 //arg_count.c
-int	check_arg_count(char **argv);
 
 #endif
